refactor of stubConfig.json + its test
