+++ wilma-application/modules/wilma-stub-configuration-parser/src/test/resources/com/epam/wilma/stubconfig/JsonWilmaStubTestValid.json
