+++ wilma-application/modules/wilma-stub-configuration-parser/src/test/resources/com/epam/wilma/stubconfig/JsonWilmaStubTestValidMinimal.json
@@ -1,0 +1,4 @@
+{
+  "wilmaStubConfiguration": {
+  }
+}
