+++ wilma-application/modules/wilma-stub-configuration-parser/src/test/resources/com/epam/wilma/stubconfig/schema/interceptors/JsonWilmaStubTestInValidInterceptorsMandatoryClassIsEmpty.json
@@ -1,8 +1,10 @@
 {
-  "interceptors": [
-    {
-      "name": "test",
-      "class": ""
-    }
-  ]
+  "wilmaStubConfiguration": {
+    "interceptors": [
+      {
+        "name": "test",
+        "class": ""
+      }
+    ]
+  }
 }
