+++ wilma-application/modules/wilma-stub-configuration-parser/src/test/resources/com/epam/wilma/stubconfig/schema/interceptors/JsonWilmaStubTestInValidInterceptorsMandatoryNameIsEmpty.json
@@ -1,8 +1,10 @@
 {
-  "interceptors": [
-    {
-      "name": "",
-      "class": "com.epam.wilma.test.stubConfig.schema"
-    }
-  ]
+  "wilmaStubConfiguration": {
+    "interceptors": [
+      {
+        "name": "",
+        "class": "com.epam.wilma.test.stubConfig.schema"
+      }
+    ]
+  }
 }
