+++ wilma-application/modules/wilma-stub-configuration-parser/src/test/resources/com/epam/wilma/stubconfig/schema/interceptors/JsonWilmaStubTestInValidInterceptorsMissingMandatoryClass.json
@@ -1,7 +1,9 @@
 {
-  "interceptors": [
-    {
-      "name": "test"
-    }
-  ]
+  "wilmaStubConfiguration": {
+    "interceptors": [
+      {
+        "name": "test"
+      }
+    ]
+  }
 }
