+++ wilma-application/modules/wilma-stub-configuration-parser/src/test/resources/com/epam/wilma/stubconfig/schema/interceptors/JsonWilmaStubTestInValidInterceptorsMissingMandatoryName.json
@@ -1,7 +1,9 @@
 {
-  "interceptors": [
-    {
-      "class": "com.epam.wilma.test.stubConfig.schema"
-    }
-  ]
+  "wilmaStubConfiguration": {
+    "interceptors": [
+      {
+        "class": "com.epam.wilma.test.stubConfig.schema"
+      }
+    ]
+  }
 }
