+++ wilma-application/modules/wilma-stub-configuration-parser/src/test/resources/com/epam/wilma/stubconfig/schema/interceptors/JsonWilmaStubTestInterceptorsValidMinimal.json
@@ -1,8 +1,10 @@
 {
-  "interceptors": [
-    {
-      "name": "a",
-      "class": "b"
-    }
-  ]
+  "wilmaStubConfiguration": {
+    "interceptors": [
+      {
+        "name": "a",
+        "class": "b"
+      }
+    ]
+  }
 }
